Remove config file, rename package
