--- sample-videochat-webrtc-new/src/main/assets/pc_config.txt
@@ -1,6 +1,0 @@
-{'iceServers': [{'url': 'stun:stun.l.google.com:19302'}, {'url': 'turn:turnserver.quickblox.com:3478?transport=udp',
-								 'username': 'user',
-								 'credential': 'user'},
-								{'url': 'turn:turnserver.quickblox.com:3478?transport=tcp',
-								  'username': 'user',
-								  'credential': 'user'}]};
